Move following logic to new reusable FollowButton and implement
