+++ frontend/src/Components/Elements/SearchThumbnails/UserThumbnail.tsx
@@ -1,5 +1,6 @@
 import { MdiDotsHorizontal } from "../../Icons/MdiThreeDots";
 import Button from "../Button";
+import FollowButton from "../Inputs/FollowButton";
 import { ProfilePicture } from "../ProfilePicture";
 
 type UserThumbnailProps = {
@@ -38,9 +39,7 @@ function UserThumbnail({
               </div>
             </div>
             <div className="flex h-fit flex-col justify-end gap-4 self-center lg:flex-row">
-              <Button className="btn-primary">
-                <small>Follow</small>
-              </Button>
+              <FollowButton followUserName={username} />
               <Button className="btn-primary w-fit self-center py-3 lg:px-3 lg:py-3">
                 <p>
                   <MdiDotsHorizontal />
